refactor: remove redux, some rearrangement
